[migration 3] Gradle 8.14.5, PMD 7.26.0, and the Phase 2 carry-overs
First of several sub-commits in Phase 3. The Boot 2.7.18 Gradle plugin configures and
builds on Gradle 8.14.5, so the toolchain move lands on its own, before the framework
move.

Gradle 7.6.6 -> 8.14.5. Two things had to change to get there:

  - GradleBuild.buildFile, set alongside dir in buildJar/buildCli/buildAll, is
    deprecated for removal in Gradle 9; Gradle's advice is to use dir, which all three
    already set to the same directory. Removed - the nested builds' task graphs are
    identical with and without it, checked on buildCli.
  - The test JVM needs a fifth --add-opens, java.base/java.lang. Gradle 8.14.5 bundles
    JaCoCo 0.8.13 where 7.6 bundled 0.8.8, and under the newer agent EhCache 2's
    reflective sizer reaches java.lang.String.value and dies there: 24 extra test
    failures with no source change. Pinning JaCoCo back hides it but is a trap - 0.8.8
    cannot read Java 21 class files. All five flags go away with Caffeine later in this
    phase.

Phase 2 carry-over: the --add-opens=java.prefs/java.util.prefs in gradle.properties is
gone. Gradle 8.14 passes it to its own daemon, so checkstyle/pmd no longer need it from
us.

PMD 6.55.0 -> 7.26.0, both modules (7.x is why Gradle 8 was needed first). Three
build-level findings:

  - io.spring.dependency-management applies the Boot BOM to every configuration,
    including the one Gradle assembles to run PMD, and Boot 2.7.18 pins commons-lang3
    below the 3.13 that PMD 7 needs - pmdMain dies with a NoSuchMethodError before
    reading a source file. Fixed with a resolutionStrategy rule on configurations.pmd,
    not a declared pmd dependency: PmdPlugin adds the analyser through
    defaultDependencies, which is skipped once the configuration has any declared
    dependency, so declaring one swaps the error for ClassNotFoundException: PMD.
  - BooleanInstantiation was removed in 7 (-> codestyle/UnnecessaryBoxing) and
    AvoidCatchingGenericException moved from design to errorprone. A stale rule
    reference fails silently: PMD logs "Cannot load ruleset" and "No files to analyze"
    and Gradle still says BUILD SUCCESSFUL.
  - Four rules this project keeps are scheduled for removal in PMD 8 and now warn on
    every run: GenericsNaming, AvoidCatchingThrowable, AvoidCatchingNPE,
    AvoidLosingExceptionInformation. None has a successor; PMD 8 is where that gets
    answered.

139 violations fixed at source - 79 catgenome:pmdMain, 33 catgenome:pmdTest,
9 ngb-cli:pmdMain, 18 ngb-cli:pmdTest. Most are self-qualifier removals, redundant
final, and unused imports. pmdTest and checkstyleTest had never run in this repo (make
lint is main-only, and make jar passes -PnoTest, which drops `check` from buildJar), so
51 of those predate PMD 7. The ones that were not mechanical:

  - FeatureIndexManager: (double) casts instead of dropping .doubleValue(), which would
    have turned double division into integer division.
  - FeatureFile: deleted six hand-written accessors that duplicate what @Getter/@Setter
    generate, rather than renaming getCompressed() to isCompressed() and leaving Lombok
    free to generate a second accessor for the same property.
  - VcfFilterForm: field isExon -> exon, so the required isExon() getter does not trip
    AvoidFieldNameMatchingMethodName. The JSON property is "exon" either way.
  - AbstractFeatureReader, AbstractEnhancedFeatureReader, FeatureIterator: exceptions no
    longer thrown into their own try's catch clause. Same types, messages and causes.
  - ProteinSequenceUtils' codon table and ngb-cli's BiologicalDataItemFormat: dropped
    fields nothing ever read. The codon table's proved it - PRO's "extended title" was
    "Ser" - and the CLI enum's numeric id was declaration order 1..22, which the ordinal
    already says.
  - util/TestUtils.assertFail keeps its catch (Exception) with a local
    @SuppressWarnings; a narrower catch cannot express what the helper is for. This is
    the exemption the deleted pmd-ruleset-feature-index-manager.xml was granting.

Verified: make jar, make lint, ngb-cli pmd/checkstyle main and test, catgenome check -x
test all green; make test 528 tests, 4 failed - the 3 documented baseline failures plus
PdbDataManagerTest.testParse, which TEST-BASELINE.md records as flapping on live RCSB
data. Findings written up in .devenv/JAVA21-MIGRATION-PLAN.md.

+++ server/catgenome/src/main/java/com/epam/catgenome/manager/protein/ProteinSequenceReconstructionManager.java
@@ -577,8 +577,8 @@ public class ProteinSequenceReconstructionManager {
             String aminoAcid = ProteinSequenceUtils.tripletToAminoAcid(
                 triple.stream().map(Sequence::getText).collect(Collectors.joining()));
 
-            long tripleStartIndex = isNegative? triple.get(2).getStartIndex().longValue()
-                                              : triple.get(0).getStartIndex().longValue();
+            long tripleStartIndex = isNegative? triple.get(2).getStartIndex()
+                                              : triple.get(0).getStartIndex();
             if (((newExtendedStart > 0) && isNegative) || ((newExtendedStart < 0) && !isNegative)) {
                 tripleStartIndex = cds.getStartIndex();
             }
